refactor: MAT-218 nameUtils, timeUtils를 공통 유틸로 통합 및 중복 제거
