--- src/routes/matches/$matchId/entry/-components/nameUtils.ts
@@ -1,7 +1,0 @@
-import { Faker, ko } from '@faker-js/faker';
-
-const faker = new Faker({
-  locale: [ko],
-});
-
-export const getRandomName = () => faker.person.fullName().slice(0, 4);
