+++ src/utils/timeUtils.ts
@@ -1,3 +1,19 @@
+const KST_OFFSET = 9 * 60 * 60 * 1000; // KST UTC+9
+const TWO_HOURS = 2 * 60 * 60 * 1000;
+
+const getKSTDate = () => new Date(Date.now() + KST_OFFSET);
+
+export const getYYYYMMDDHHMMSS = () =>
+  getKSTDate()
+    .toISOString()
+    .replace(/[-:T.Z]/g, '')
+    .slice(0, 14);
+
+export const YYYYMMDD = () => getKSTDate().toISOString().slice(0, 10);
+export const nowHHMMSS = () => getKSTDate().toISOString().slice(11, 19);
+export const twoHoursLaterHHMMSS = () =>
+  new Date(Date.now() + KST_OFFSET + TWO_HOURS).toISOString().slice(11, 19);
+
 // NOTE: mock용으로 생성했던 함수여서 앞으로는 필요 없을 듯
 export const getTimeAgo = ({
   minutes = 0,
